- Move credit card functions to static rather than instance - Move credit card unit tests from PaymentMethodTest to CreditCardTest
